+++ lib/lottie.dart
@@ -2,4 +2,5 @@ export 'src/composition.dart' show LottieComposition;
 export 'src/lottie.dart' show Lottie;
 export 'src/lottie_builder.dart' show LottieBuilder;
 export 'src/lottie_drawable.dart' show LottieDrawable;
+export 'src/lottie_image_asset.dart' show LottieImageAsset;
 export 'src/raw_lottie.dart' show RawLottie;